Split `compile` integration tests in two parts
